+++ docs/guides/changelog.md
@@ -10,6 +10,7 @@ description: |-
 ## Unreleased
 
 - (Enh) Device interfaces data sources now accept `logical_name` as search parameter.
+- (Enh) If `items` in bulk data source is empty, all objects from FMC will be loaded
 
 ## 2.1.0